--- OPP_Projects/HWsolution.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.9.34607.119
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Ctime", "Ctime\Ctime.vcxproj", "{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Debug|x64.ActiveCfg = Debug|x64
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Debug|x64.Build.0 = Debug|x64
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Debug|x86.ActiveCfg = Debug|Win32
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Debug|x86.Build.0 = Debug|Win32
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Release|x64.ActiveCfg = Release|x64
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Release|x64.Build.0 = Release|x64
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Release|x86.ActiveCfg = Release|Win32
-		{F9C40BD9-2715-4541-8B1C-C23F19DBB9D0}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4C982206-4FCF-4DBB-A9B5-D620083CC5B5}
-	EndGlobalSection
-EndGlobal
