+++ OPP_Projects/OppSolutions.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34607.119
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Ctime", "Ctime\Ctime.vcxproj", "{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Debug|x64.ActiveCfg = Debug|x64
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Debug|x64.Build.0 = Debug|x64
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Debug|x86.ActiveCfg = Debug|Win32
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Debug|x86.Build.0 = Debug|Win32
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Release|x64.ActiveCfg = Release|x64
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Release|x64.Build.0 = Release|x64
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Release|x86.ActiveCfg = Release|Win32
+		{9916DB30-7DC9-4AF6-99A6-BCD56AADBB12}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {EA0544D0-8DD9-49DD-B551-98CF560982A2}
+	EndGlobalSection
+EndGlobal
